+++ Editor/SceneTemplatePipeline.cs
@@ -10,6 +10,12 @@ using UnityEngine.SceneManagement;
 public interface ISceneTemplatePipeline
 {
     /// <summary>
+    /// Event called before we display the New Template Dialog to determine if this template should be available to user.
+    /// </summary>
+    /// <param name="sceneTemplateAsset">Scene Template asset that would be avauilable in New Scene Dialog</param>    
+    bool IsValidTemplateForInstantiation(SceneTemplateAsset sceneTemplateAsset);
+
+    /// <summary>
     /// Event called before the template is instantiated.
     /// </summary>
     /// <param name="sceneTemplateAsset">Scene Template asset that will be instantiated</param>
@@ -32,14 +38,24 @@ public interface ISceneTemplatePipeline
 public class SceneTemplatePipelineAdapter : ISceneTemplatePipeline
 {
     /// <summary>
+    /// Event called before we display the New Template Dialog to determine if this template should be available to user.
+    /// </summary>
+    /// <param name="sceneTemplateAsset">Scene Template asset that would be avauilable in New Scene Dialog</param>    
+    public virtual bool IsValidTemplateForInstantiation(SceneTemplateAsset sceneTemplateAsset)
+    {
+        return true;
+    }
+
+    /// <summary>
     /// Event called before the template is instantiated.
     /// </summary>
     /// <param name="sceneTemplateAsset">Scene Template asset that will be instantiated</param>
     /// <param name="isAdditive">Is the new scene created in additive mode.</param>
     /// <param name="sceneName">What is the path of the new scene (could be empty if untitled)</param>
-    public void BeforeTemplateInstantiation(SceneTemplateAsset sceneTemplateAsset, bool isAdditive, string sceneName)
+    public virtual void BeforeTemplateInstantiation(SceneTemplateAsset sceneTemplateAsset, bool isAdditive, string sceneName)
     {
     }
+
     /// <summary>
     /// Event called after the template is instantiated and the new scene is still loaded.
     /// </summary>
@@ -47,7 +63,7 @@ public class SceneTemplatePipelineAdapter : ISceneTemplatePipeline
     /// <param name="scene">The newly created scene</param>
     /// <param name="isAdditive">Is the new scene created in additive mode.</param>
     /// <param name="sceneName">What is the path of the new scene (could be empty if untitled)</param>
-    public void AfterTemplateInstantiation(SceneTemplateAsset sceneTemplateAsset, Scene scene, bool isAdditive, string sceneName)
+    public virtual void AfterTemplateInstantiation(SceneTemplateAsset sceneTemplateAsset, Scene scene, bool isAdditive, string sceneName)
     {
     }
 }