+++ Tests/Editor/HideSceneTemplatePipeline.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace UnityEditor.SceneTemplate
+{
+    public class HideSceneTemplatePipeline : SceneTemplatePipelineAdapter
+    {
+        public override bool IsValidTemplateForInstantiation(SceneTemplateAsset sceneTemplateAsset)
+        {
+            return false;
+        }
+    }
+}